content/canvas/test/test_2d.path.arc.negative.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.arc.negative</title>
<!-- Testing: arc() with negative radius throws INDEX_SIZE_ERR -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown = undefined; try {
  ctx.arc(0, 0, -1, 0, 0, true);
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.INDEX_SIZE_ERR, "should throw INDEX_SIZE_ERR");

SimpleTest.finish();

});
</script>
